fix missing function
