+++ app/init/app.php
@@ -45,6 +45,14 @@ use Utopia\Pools\Group;
 use Utopia\Pools\Pool;
 use Utopia\Queue;
 use Utopia\Registry\Registry;
+use Utopia\Storage\Device;
+use Utopia\Storage\Device\Backblaze;
+use Utopia\Storage\Device\DOSpaces;
+use Utopia\Storage\Device\Linode;
+use Utopia\Storage\Device\Local;
+use Utopia\Storage\Device\S3;
+use Utopia\Storage\Device\Wasabi;
+use Utopia\Storage\Storage;
 use Utopia\System\System;
 use Utopia\Validator\IP;
 use Utopia\Validator\Range;
@@ -1041,3 +1049,90 @@ foreach ($locales as $locale) {
         'timeout' => 2,
     ],
 ]);
+
+
+function getDevice($root): Device
+{
+    $connection = System::getEnv('_APP_CONNECTIONS_STORAGE', '');
+
+    if (!empty($connection)) {
+        $acl = 'private';
+        $device = Storage::DEVICE_LOCAL;
+        $accessKey = '';
+        $accessSecret = '';
+        $bucket = '';
+        $region = '';
+
+        try {
+            $dsn = new DSN($connection);
+            $device = $dsn->getScheme();
+            $accessKey = $dsn->getUser() ?? '';
+            $accessSecret = $dsn->getPassword() ?? '';
+            $bucket = $dsn->getPath() ?? '';
+            $region = $dsn->getParam('region');
+        } catch (\Throwable $e) {
+            Console::warning($e->getMessage() . 'Invalid DSN. Defaulting to Local device.');
+        }
+
+        switch ($device) {
+            case Storage::DEVICE_S3:
+                return new S3($root, $accessKey, $accessSecret, $bucket, $region, $acl);
+            case STORAGE::DEVICE_DO_SPACES:
+                $device = new DOSpaces($root, $accessKey, $accessSecret, $bucket, $region, $acl);
+                $device->setHttpVersion(S3::HTTP_VERSION_1_1);
+                return $device;
+            case Storage::DEVICE_BACKBLAZE:
+                return new Backblaze($root, $accessKey, $accessSecret, $bucket, $region, $acl);
+            case Storage::DEVICE_LINODE:
+                return new Linode($root, $accessKey, $accessSecret, $bucket, $region, $acl);
+            case Storage::DEVICE_WASABI:
+                return new Wasabi($root, $accessKey, $accessSecret, $bucket, $region, $acl);
+            case Storage::DEVICE_LOCAL:
+            default:
+                return new Local($root);
+        }
+    } else {
+        switch (strtolower(System::getEnv('_APP_STORAGE_DEVICE', Storage::DEVICE_LOCAL) ?? '')) {
+            case Storage::DEVICE_LOCAL:
+            default:
+                return new Local($root);
+            case Storage::DEVICE_S3:
+                $s3AccessKey = System::getEnv('_APP_STORAGE_S3_ACCESS_KEY', '');
+                $s3SecretKey = System::getEnv('_APP_STORAGE_S3_SECRET', '');
+                $s3Region = System::getEnv('_APP_STORAGE_S3_REGION', '');
+                $s3Bucket = System::getEnv('_APP_STORAGE_S3_BUCKET', '');
+                $s3Acl = 'private';
+                return new S3($root, $s3AccessKey, $s3SecretKey, $s3Bucket, $s3Region, $s3Acl);
+            case Storage::DEVICE_DO_SPACES:
+                $doSpacesAccessKey = System::getEnv('_APP_STORAGE_DO_SPACES_ACCESS_KEY', '');
+                $doSpacesSecretKey = System::getEnv('_APP_STORAGE_DO_SPACES_SECRET', '');
+                $doSpacesRegion = System::getEnv('_APP_STORAGE_DO_SPACES_REGION', '');
+                $doSpacesBucket = System::getEnv('_APP_STORAGE_DO_SPACES_BUCKET', '');
+                $doSpacesAcl = 'private';
+                $device = new DOSpaces($root, $doSpacesAccessKey, $doSpacesSecretKey, $doSpacesBucket, $doSpacesRegion, $doSpacesAcl);
+                $device->setHttpVersion(S3::HTTP_VERSION_1_1);
+                return $device;
+            case Storage::DEVICE_BACKBLAZE:
+                $backblazeAccessKey = System::getEnv('_APP_STORAGE_BACKBLAZE_ACCESS_KEY', '');
+                $backblazeSecretKey = System::getEnv('_APP_STORAGE_BACKBLAZE_SECRET', '');
+                $backblazeRegion = System::getEnv('_APP_STORAGE_BACKBLAZE_REGION', '');
+                $backblazeBucket = System::getEnv('_APP_STORAGE_BACKBLAZE_BUCKET', '');
+                $backblazeAcl = 'private';
+                return new Backblaze($root, $backblazeAccessKey, $backblazeSecretKey, $backblazeBucket, $backblazeRegion, $backblazeAcl);
+            case Storage::DEVICE_LINODE:
+                $linodeAccessKey = System::getEnv('_APP_STORAGE_LINODE_ACCESS_KEY', '');
+                $linodeSecretKey = System::getEnv('_APP_STORAGE_LINODE_SECRET', '');
+                $linodeRegion = System::getEnv('_APP_STORAGE_LINODE_REGION', '');
+                $linodeBucket = System::getEnv('_APP_STORAGE_LINODE_BUCKET', '');
+                $linodeAcl = 'private';
+                return new Linode($root, $linodeAccessKey, $linodeSecretKey, $linodeBucket, $linodeRegion, $linodeAcl);
+            case Storage::DEVICE_WASABI:
+                $wasabiAccessKey = System::getEnv('_APP_STORAGE_WASABI_ACCESS_KEY', '');
+                $wasabiSecretKey = System::getEnv('_APP_STORAGE_WASABI_SECRET', '');
+                $wasabiRegion = System::getEnv('_APP_STORAGE_WASABI_REGION', '');
+                $wasabiBucket = System::getEnv('_APP_STORAGE_WASABI_BUCKET', '');
+                $wasabiAcl = 'private';
+                return new Wasabi($root, $wasabiAccessKey, $wasabiSecretKey, $wasabiBucket, $wasabiRegion, $wasabiAcl);
+        }
+    }
+}
